Unit tests for settings screen.
