+++ src/lib/Scenes/Settings/__mocks__/LoggedInUserInfo.tsx
@@ -1,0 +1,6 @@
+import { Serif } from "@artsy/palette"
+import React from "react"
+
+export const UserProfileQueryRenderer: React.FC = () => (
+  <Serif size="3t">Logged in as: Unit Test User (example@example.com)</Serif>
+)
